Initial values were adjusted.

+++ src/ukf.h
@@ -6,6 +6,7 @@
 #include <vector>
 #include <string>
 #include <fstream>
+#include "tools.h"
 
 using Eigen::MatrixXd;
 using Eigen::VectorXd;
@@ -65,14 +66,13 @@ public:
   int n_aug_;
 
   ///* Sigma point spreading parameter
-  int lambda_;
+  double lambda_;
 
   ///* the current NIS for radar
   double NIS_radar_;
 
   ///* the current NIS for laser
   double NIS_laser_;
-
 
   /**
    * Constructor
